Slowly modularizing methods - finished events and dataMatch

+++ Event.java
@@ -4,6 +4,8 @@ import java.io.FileWriter;
 
 //Class for creating a singular event with instances like title, description, location, time, member list, and date.
 public class Event {
+
+    private static DataMatch eventComparator = new DataMatch();
     private String eventTitle;
     private String eventDescription;
     private String eventLocation;
@@ -53,28 +55,30 @@ public class Event {
         return attendees.size();
     }
 
-    //When the event is printed, it will display all the details of the event.
-    public String toFileString() {
-        return eventID + "," + eventTitle + "," + eventLocation + "," + eventTime + "," + eventDate + "\n";
-    }
+    //May need to use later 
+    // private String toFileString() {
+    //     return eventID + "," + eventTitle + "," + eventLocation + "," + eventTime + "," + eventDate + "\n";
+    // }
 
+    //When the event is printed, it will display all the details of the event.
     public String toString() {
         return "\nEvent ID: " + eventID + "\nEvent Title: " + eventTitle + "\nDescription: " + eventDescription +
                 "\nLocation: " + eventLocation + "\nTime: " + eventTime + "\nDate: " + eventDate + "\n";   
     }
 
-    public boolean doesEventExistInFile(String input, int index) {
-        DataMatch fileSearcher = new DataMatch();
-        return fileSearcher.isDataInFile(input, index);
+    private boolean doesEventExistInFile() {
+        return eventComparator.isDataInFile("" + eventID, 0, "EventList.txt");
     }
 
     //Write into file- all details of the event.
     public void addEventToFile() {
-        try(BufferedWriter writer = new BufferedWriter(new FileWriter("EventList.txt", true))) {
-            writer.write(eventID + "," + eventTitle + "," + eventLocation + "," + eventTime + "," + eventDate + "\n");
-        }
-        catch(Exception e) {
-            e.printStackTrace();
+        if(doesEventExistInFile() == false) {
+            try(BufferedWriter writer = new BufferedWriter(new FileWriter("EventList.txt", true))) {
+                writer.write(eventID + "," + eventTitle + "," + eventLocation + "," + eventTime + "," + eventDate + "\n");
+            }
+            catch(Exception e) {
+                e.printStackTrace();
+            }
         }
     }
